level system change 60

+++ include/limit.h
@@ -12,5 +12,5 @@ struct title_type
 {
 	char *title_m;
 	char *title_f;
-	int exp;
+	long int exp;
 };
